% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFCP.R
\name{BFCP}
\alias{BFCP}
\title{A Bimodal fitting calculation parallel (BFCP)computing function}
\usage{
BFCP(mat, parallel = TRUE)
}
\arguments{
\item{mat}{Input a data frame}
}
\description{
A Bimodal fitting calculation parallel (BFCP)computing function
}
\examples{
brca <- BFCP((datL_brca_all$mat), parallel=TRUE)
}
\keyword{Bimodal}
