% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFCP.R
\name{loadPDUIdata}
\alias{loadPDUIdata}
\title{Load PUDI txt file}
\usage{
loadPDUIdata(file, col_anno = 1:3, header = TRUE, TvsN = FALSE)
}
\arguments{
\item{file}{txt file}
}
\description{
Load PUDI txt file
}
\examples{
datL_brca <- loadPDUIdata(file=file.path(DirInputData, 'BRCA_known_APA_Combined_PDUIs.txt'),col_anno=1:3,header=T)
}
\keyword{PUDI}
\keyword{load}
