% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFCP.R
\name{plotGene}
\alias{plotGene}
\title{plot normal distribution mix for one gene_id}
\usage{
plotGene(dataraw, DI, gene)
}
\arguments{
\item{dataraw}{raw dataframe}

\item{DI}{calculated Di value dataframe}

\item{gene}{gene id name}
}
\description{
plot normal distribution mix for one gene_id
}
\examples{
plotGene(datL_brca_all,di_brca_all,gene)
}
